Move second-quantization-specific methods from KroneckerDelta to a secondquant subclass
The general-purpose KroneckerDelta in
sympy/functions/special/tensor_functions.py carried methods that only make
sense for the second quantization machinery (is_above_fermi, is_below_fermi,
is_only_above_fermi, is_only_below_fermi, indices_contain_equal_information,
preferred_index, killable_index and the _get_preferred_index helper).

These methods rely on the above_fermi/below_fermi assumptions and are used
exclusively by sympy/physics/secondquant.py. Move them onto a KroneckerDelta
subclass defined in secondquant.py, keeping the general-purpose class free of
second-quantization-specific logic. The core eval/symmetry/canonicalization
logic and delta_range (consumed by sympy/concrete/delta.py) stay on the base
class.

secondquant now exposes the subclass under the name KroneckerDelta and builds
it for every delta it creates, so public behavior is unchanged. evaluate_deltas
promotes any plain KroneckerDelta it receives to the subclass so that callers
passing the general-purpose delta keep working. The fermi-specific tests move
from test_tensor_functions.py to test_secondquant.py.

Fixes #6271.
